fix: executiontabcontent
