+++ packages/zapp/console/src/components/Executions/ExecutionDetails/Timeline/ExecutionTimeline.tsx
@@ -1,17 +1,11 @@
 import * as React from 'react';
 import { makeStyles, Typography } from '@material-ui/core';
-
-import { useNodeExecutionContext } from 'components/Executions/contextProvider/NodeExecutionDetails';
-import { transformerWorkflowToDag } from 'components/WorkflowGraph/transformerWorkflowToDag';
 import { isEndNode, isStartNode, isExpanded } from 'components/WorkflowGraph/utils';
 import { tableHeaderColor } from 'components/Theme/constants';
 import { timestampToDate } from 'common/utils';
 import { dNode } from 'models/Graph/types';
-import { makeNodeExecutionDynamicWorkflowQuery } from 'components/Workflow/workflowQueries';
-import { useQuery } from 'react-query';
 import { createRef, useContext, useEffect, useRef, useState } from 'react';
 import { NodeExecutionsByIdContext } from 'components/Executions/contexts';
-import { checkForDynamicExecutions } from 'components/common/utils';
 import { convertToPlainNodes } from './helpers';
 import { ChartHeader } from './ChartHeader';
 import { useScaleContext } from './scaleContext';
@@ -72,9 +66,10 @@ const INTERVAL_LENGTH = 110;
 
 interface ExProps {
   chartTimezone: string;
+  initializeNodes: dNode[];
 }
 
-export const ExecutionTimeline: React.FC<ExProps> = ({ chartTimezone }) => {
+export const ExecutionTimeline: React.FC<ExProps> = ({ chartTimezone, initializeNodes }) => {
   const [chartWidth, setChartWidth] = useState(0);
   const [labelInterval, setLabelInterval] = useState(INTERVAL_LENGTH);
   const durationsRef = useRef<HTMLDivElement>(null);
@@ -84,29 +79,12 @@ export const ExecutionTimeline: React.FC<ExProps> = ({ chartTimezone }) => {
   const [originalNodes, setOriginalNodes] = useState<dNode[]>([]);
   const [showNodes, setShowNodes] = useState<dNode[]>([]);
   const [startedAt, setStartedAt] = useState<Date>(new Date());
-
-  const { compiledWorkflowClosure } = useNodeExecutionContext();
-  const { chartInterval: chartTimeInterval } = useScaleContext();
-  const { staticExecutionIdsMap } = compiledWorkflowClosure
-    ? transformerWorkflowToDag(compiledWorkflowClosure)
-    : { staticExecutionIdsMap: {} };
-
   const nodeExecutionsById = useContext(NodeExecutionsByIdContext);
-
-  const dynamicParents = checkForDynamicExecutions(nodeExecutionsById, staticExecutionIdsMap);
-
-  const { data: dynamicWorkflows } = useQuery(
-    makeNodeExecutionDynamicWorkflowQuery(dynamicParents),
-  );
+  const { chartInterval: chartTimeInterval } = useScaleContext();
 
   useEffect(() => {
-    const nodes: dNode[] = compiledWorkflowClosure
-      ? transformerWorkflowToDag(compiledWorkflowClosure, dynamicWorkflows).dag.nodes
-      : [];
-    // we remove start/end node info in the root dNode list during first assignment
-    const initializeNodes = convertToPlainNodes(nodes);
     setOriginalNodes(initializeNodes);
-  }, [dynamicWorkflows, compiledWorkflowClosure]);
+  }, [initializeNodes]);
 
   useEffect(() => {
     const initializeNodes = convertToPlainNodes(originalNodes);
